kernel: cutedsl backend support D [320,512)
